requirements.txt
terminaltables==3.1.0
python-dotenv==0.10.3
python-telegram-bot==12.2.0
requests==2.22.0
urllib3==1.25.7
